add convgru
